Hscaffold

Running Hscaffold

runHscaffold :: MonadIO m => FilePath -> WriterT ScaffoldActionV m a -> m a

Run the scaffolding writer on the IO monad with no extensions

runHscaffold "." $ do
    file "./.gitignore" (Text.unlines [ ".stack-work"
                                      , "stuff"
                                      , "here"
                                      ])
    directory "./src" $ do
        file "./Main.hs" "main = putStrLn \"Hello World\""
        file "./Other.hs" "other = putStrLn \"Hello You\""

Convert Hscaffold to HSFILES (stack templates)

toHsfiles :: Writer ScaffoldActionV a -> Text

Run the scaffolding writer and return an .hsfiles Text to use with stack-templates

writeToHsfiles :: MonadIO m => FilePath -> Writer ScaffoldActionV a -> m ()

Shortcut for

writeToHsfiles = do h <- toHsfiles w; liftIO $ Text.writeFile fp h

Convert HSFILES to Hscaffold

fromHsfiles :: String -> ScaffoldAction e

Converts hsfiles to hscaffold actions

fromHsfilesW :: MonadWriter (ScaffoldAction e) m => String -> m ()

Converts hsfiles to a hscaffold monad

readHsfiles :: FilePath -> IO (ScaffoldAction e)

Convert a directory to Hscaffold

hscaffoldFromDirectory :: FilePath -> IO (ScaffoldAction e)

Converts a directory to scaffold actions

Compile Hscaffold to Haskell code

hscaffoldToHaskell :: Foldable t => t (ScaffoldActionType e) -> Text

Generates Haskell code from scaffold actions

Finer grained runners

runAction :: FilePath -> ScaffoldActionType e -> IO ()

Run a single scaffolding action on the IO monad with no extensions

runWriter :: Writer w a -> (a, w)

Unwrap a writer computation as a (result, output) pair. (The inverse of writer.)

runWriterT :: WriterT w m a -> m (a, w)

EDSL Combinators

directory :: MonadWriter (ScaffoldAction e) m => FilePath -> WriterT (ScaffoldAction e) m b -> m b

Create a directory with the nested contents

file :: MonadWriter (ScaffoldAction e) m => FilePath -> Text -> m ()

Create a file with the given contents

link :: MonadWriter (ScaffoldAction e) m => FilePath -> FilePath -> m ()

Create a symbolic link between two filepaths

copy :: MonadWriter (ScaffoldAction e) m => FilePath -> FilePath -> m ()

Copy a file from A to B

Non-absolute paths are treated relative to the *current* root, nested blocks change the root

touch :: MonadWriter (ScaffoldAction e) m => FilePath -> m ()

Write the empty string to a file

Setting permissions

permissions :: MonadWriter (ScaffoldAction e) m => FilePath -> Permissions -> m ()

Set permissions on a filepath

fileWith :: MonadWriter (ScaffoldAction e) m => Permissions -> FilePath -> Text -> m ()

Create a file with the given contents and permissions

directoryWith :: MonadWriter (ScaffoldAction e) m => Permissions -> FilePath -> WriterT (ScaffoldAction e) m b -> m b

Create a directory with the nested contents and permissions

copyWith :: MonadWriter (ScaffoldAction e) m => Permissions -> FilePath -> FilePath -> m ()

Copy a file from A to B and set permissions on B, see copy

touchWith :: MonadWriter (ScaffoldAction e) m => Permissions -> FilePath -> m ()

Write the empty string to a file with given permissions

data Permissions :: *

Types

type ScaffoldMonadT m a = WriterT ScaffoldActionV m a

The writer monad transformer for scaffold actions

type ScaffoldMonadIO a = WriterT ScaffoldActionV IO a

The writer monad for scaffold actions, running in IO

data ScaffoldActionType e

Type of actions scaffolding can make, ScaffoldActionTypeExtension is open for extension through other data-types

Constructors

File FilePath Text

Link FilePath FilePath

Directory FilePath (ScaffoldAction e)

SetPermissions Permissions FilePath

Copy FilePath FilePath

ScaffoldActionTypeExtension e

type ScaffoldAction e = [ScaffoldActionType e]

Accumulator for actions

type ScaffoldActionV = ScaffoldAction ()

Accumulator for actions set with void extension

type ScaffoldMonadET e m a = WriterT (ScaffoldAction e) m a

The writer monad transformer for scaffold actions with an extension

Helpers

runHscaffoldIO :: FilePath -> ScaffoldMonadIO a -> IO a

mkActionPath

Arguments

:: FilePath
The root

-> FilePath
A filepath

-> FilePath

Very simple helper, if the second argument is absolute, returns it, otherwise, return it relative to the first argument

Re-exports

data Text :: *

module Control.Monad.IO.Class

module Control.Monad.Writer

module System.Directory

module System.FilePath