Hscaffold.EDSL

Documentation

directory :: MonadWriter (ScaffoldAction e) m => FilePath -> WriterT (ScaffoldAction e) m b -> m b

Create a directory with the nested contents

directoryWith :: MonadWriter (ScaffoldAction e) m => Permissions -> FilePath -> WriterT (ScaffoldAction e) m b -> m b

Create a directory with the nested contents and permissions

file :: MonadWriter (ScaffoldAction e) m => FilePath -> Text -> m ()

Create a file with the given contents

fileWith :: MonadWriter (ScaffoldAction e) m => Permissions -> FilePath -> Text -> m ()

Create a file with the given contents and permissions

permissions :: MonadWriter (ScaffoldAction e) m => FilePath -> Permissions -> m ()

Set permissions on a filepath

link :: MonadWriter (ScaffoldAction e) m => FilePath -> FilePath -> m ()

Create a symbolic link between two filepaths

touch :: MonadWriter (ScaffoldAction e) m => FilePath -> m ()

Write the empty string to a file

touchWith :: MonadWriter (ScaffoldAction e) m => Permissions -> FilePath -> m ()

Write the empty string to a file with given permissions

copy :: MonadWriter (ScaffoldAction e) m => FilePath -> FilePath -> m ()

Copy a file from A to B

Non-absolute paths are treated relative to the *current* root, nested blocks change the root

copyWith :: MonadWriter (ScaffoldAction e) m => Permissions -> FilePath -> FilePath -> m ()

Copy a file from A to B and set permissions on B, see copy