Hscaffold.Interpreter.Common

Documentation

mkActionPath

Arguments

:: FilePath
The root

-> FilePath
A filepath

-> FilePath

Very simple helper, if the second argument is absolute, returns it, otherwise, return it relative to the first argument