Hscaffold.Types

Documentation

type ScaffoldMonadT m a = WriterT ScaffoldActionV m a

The writer monad transformer for scaffold actions

type ScaffoldMonadIO a = WriterT ScaffoldActionV IO a

The writer monad for scaffold actions, running in IO

type ScaffoldMonadET e m a = WriterT (ScaffoldAction e) m a

The writer monad transformer for scaffold actions with an extension

type ScaffoldAction e = [ScaffoldActionType e]

Accumulator for actions

type ScaffoldActionV = ScaffoldAction ()

Accumulator for actions set with void extension

data ScaffoldActionType e

Type of actions scaffolding can make, ScaffoldActionTypeExtension is open for extension through other data-types

Constructors

File FilePath Text

Link FilePath FilePath

Directory FilePath (ScaffoldAction e)

SetPermissions Permissions FilePath

Copy FilePath FilePath

ScaffoldActionTypeExtension e